fit for mysql/mariadb authorization methods

+++ libraries/common.lib.php
@@ -3778,8 +3778,8 @@ function PMA_analyzeLimitClause($limit_clause)
 {
     $start_and_length = explode(',', str_ireplace('LIMIT', '', $limit_clause));
     return array(
-        'start'  => trim($start_and_length[0]),
-        'length' => trim($start_and_length[1])
+        'start'  => (int)trim($start_and_length[0]),
+        'length' => (int)trim($start_and_length[1])
     );
 }
